This commit was generated by cvs2svn to compensate for changes in r11871, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@11872 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/resources.c
@@ -266,7 +266,7 @@ int resources_init(const char *machine)
        pointers into the array because the array may be reallocated. */
     hashTable = (int *)xmalloc((1 << logHashSize) * sizeof(int));
 
-    for (i = 0; i < (1 << logHashSize); i++)
+    for (i = 0; i < (unsigned int)(1 << logHashSize); i++)
         hashTable[i] = -1;
 
     return 0;
